(Sylvia)SignalsModule binds NetworkTurnInfoBuilder to NetworkWithSignalsTurnInfoBuilder if linkToLinkRoutingEnabled
