+++ matsim/src/test/java/org/matsim/integration/invertednetworks/InvertedNetworkRoutingIT.java
@@ -26,7 +26,6 @@ import org.junit.Test;
 import org.matsim.core.controler.Controler;
 import org.matsim.core.controler.events.StartupEvent;
 import org.matsim.core.controler.listener.StartupListener;
-import org.matsim.core.router.LinkToLinkRoutingGuiceModule;
 import org.matsim.testcases.MatsimTestUtils;
 
 public class InvertedNetworkRoutingIT {
@@ -39,7 +38,6 @@ public class InvertedNetworkRoutingIT {
 		InvertedNetworkRoutingTestFixture f = new InvertedNetworkRoutingTestFixture(false, true, false);
 		f.scenario.getConfig().controler().setOutputDirectory(testUtils.getOutputDirectory());
 		Controler c = new Controler(f.scenario);
-		c.addOverridingModule(new LinkToLinkRoutingGuiceModule());
 		c.getConfig().controler().setDumpDataAtEnd(false);
 		c.getConfig().controler().setCreateGraphs(false);
         final InvertedNetworkRoutingTestEventHandler testHandler = new InvertedNetworkRoutingTestEventHandler();
@@ -59,7 +57,6 @@ public class InvertedNetworkRoutingIT {
 		InvertedNetworkRoutingTestFixture f = new InvertedNetworkRoutingTestFixture(true, false, false);
 		f.scenario.getConfig().controler().setOutputDirectory(testUtils.getOutputDirectory());
 		Controler c = new Controler(f.scenario);
-		c.addOverridingModule(new LinkToLinkRoutingGuiceModule());
 		c.getConfig().controler().setDumpDataAtEnd(false);
 		c.getConfig().controler().setCreateGraphs(false);
         final InvertedNetworkRoutingTestEventHandler testHandler = new InvertedNetworkRoutingTestEventHandler();
