+++ matsim/src/test/java/org/matsim/integration/invertednetworks/LanesIT.java
@@ -43,7 +43,6 @@ import org.matsim.core.mobsim.framework.listeners.MobsimInitializedListener;
 import org.matsim.core.mobsim.qsim.QSim;
 import org.matsim.core.mobsim.qsim.interfaces.NetsimLink;
 import org.matsim.core.mobsim.qsim.qnetsimengine.QLinkLanesImpl;
-import org.matsim.core.router.LinkToLinkRoutingGuiceModule;
 import org.matsim.testcases.MatsimTestUtils;
 
 /**
@@ -72,7 +71,6 @@ public class LanesIT {
 		config.controler().setLastIteration(lastIteration);
 		// ---
 		Controler controler = new Controler(config);
-		controler.addOverridingModule(new LinkToLinkRoutingGuiceModule());
 		controler.getConfig().controler().setCreateGraphs(false);
 		controler.addOverridingModule(new AbstractModule() {
 			@Override
