+++ playgrounds/michalm/src/main/java/playground/michalm/poznan/PoznanControler.java
@@ -2,7 +2,6 @@ package playground.michalm.poznan;
 
 import org.matsim.core.config.*;
 import org.matsim.core.controler.Controler;
-import org.matsim.core.router.LinkToLinkRoutingGuiceModule;
 
 
 public class PoznanControler
@@ -16,10 +15,6 @@ public class PoznanControler
 
         //MatsimRandom.reset(43);
         Controler controler = new Controler(config);
-        if (config.controler().isLinkToLinkRoutingEnabled()) {
-            controler.addOverridingModule(new LinkToLinkRoutingGuiceModule());
-        }
-
         //controler.addOverridingModule(new OTFVisLiveModule());
 
         controler.run();
