+++ playgrounds/stockholm/src/main/java/gunnar/ihop2/transmodeler/run/TransmodelerMATSim.java
@@ -12,7 +12,6 @@ import org.matsim.core.config.ConfigUtils;
 import org.matsim.core.controler.AbstractModule;
 import org.matsim.core.controler.Controler;
 import org.matsim.core.mobsim.framework.Mobsim;
-import org.matsim.core.router.LinkToLinkRoutingGuiceModule;
 
 import cadyts.utilities.misc.StreamFlushHandler;
 import floetteroed.utilities.SimpleLogFormatter;
@@ -217,7 +216,9 @@ public class TransmodelerMATSim {
 				"Starting Assignment ...");
 
 		final Controler controler = new Controler(matsimConfig);
-		controler.addOverridingModule(new LinkToLinkRoutingGuiceModule());
+        //LinkToLinkRouting is added automatically when config.controler.linkToLinkRoutingEnabled == true
+		//michalm, jan'17
+		//controler.addOverridingModule(new LinkToLinkRoutingGuiceModule());
 
 		if (MOBSIM.transmodeler.equals(mobsim)) {
