+++ playgrounds/tthunig/src/main/java/scenarios/cottbus/run/TtRunCottbusSimulation.java
@@ -35,7 +35,6 @@ import org.matsim.contrib.signals.data.SignalsDataLoader;
 import org.matsim.contrib.signals.data.signalcontrol.v20.SignalControlWriter20;
 import org.matsim.contrib.signals.data.signalgroups.v20.SignalGroupsWriter20;
 import org.matsim.contrib.signals.data.signalsystems.v20.SignalSystemsWriter20;
-import org.matsim.contrib.signals.router.LinkToLinkRoutingWithSignalsDataGuiceModule;
 import org.matsim.core.config.Config;
 import org.matsim.core.config.ConfigUtils;
 import org.matsim.core.config.ConfigWriter;
@@ -379,11 +378,6 @@ public class TtRunCottbusSimulation {
 			controler.addOverridingModule(new SignalsModule());
 		}
 		
-		// add the module for link to link routing if enabled
-		if (config.controler().isLinkToLinkRoutingEnabled()){
-			controler.addOverridingModule(new LinkToLinkRoutingWithSignalsDataGuiceModule());
-		}
-
 		if (!PRICING_TYPE.equals(PricingType.NONE) && !PRICING_TYPE.equals(PricingType.FLOWBASED)){
 			// add tolling
 			TollHandler tollHandler = new TollHandler(scenario);
