+++ playgrounds/tthunig/src/main/java/scenarios/illustrative/braess/run/RunBraessSimulation.java
@@ -40,7 +40,6 @@ import org.matsim.contrib.signals.data.SignalsDataLoader;
 import org.matsim.contrib.signals.data.signalcontrol.v20.SignalControlWriter20;
 import org.matsim.contrib.signals.data.signalgroups.v20.SignalGroupsWriter20;
 import org.matsim.contrib.signals.data.signalsystems.v20.SignalSystemsWriter20;
-import org.matsim.contrib.signals.router.LinkToLinkRoutingWithSignalsDataGuiceModule;
 import org.matsim.core.config.Config;
 import org.matsim.core.config.ConfigUtils;
 import org.matsim.core.config.ConfigWriter;
@@ -314,11 +313,6 @@ public final class RunBraessSimulation {
 			});
 		}
 		
-		// add the module for link to link routing if enabled
-		if (config.controler().isLinkToLinkRoutingEnabled()){
-			controler.addOverridingModule(new LinkToLinkRoutingWithSignalsDataGuiceModule());
-		}
-
 		if (!PRICING_TYPE.equals(PricingType.NONE) && !PRICING_TYPE.equals(PricingType.FLOWBASED) && !PRICING_TYPE.equals(PricingType.GREGOR) && !PRICING_TYPE.equals(PricingType.INTERVALBASED)){
 			// add tolling
 			TollHandler tollHandler = new TollHandler(scenario);
