+++ playgrounds/tthunig/src/main/java/utils/TtBasicController.java
@@ -24,7 +24,6 @@ import org.matsim.contrib.signals.SignalSystemsConfigGroup;
 import org.matsim.contrib.signals.controler.SignalsModule;
 import org.matsim.contrib.signals.data.SignalsData;
 import org.matsim.contrib.signals.data.SignalsDataLoader;
-import org.matsim.contrib.signals.router.LinkToLinkRoutingWithSignalsDataGuiceModule;
 import org.matsim.core.config.Config;
 import org.matsim.core.config.ConfigUtils;
 import org.matsim.core.controler.Controler;
@@ -61,11 +60,6 @@ public class TtBasicController {
 			controler.addOverridingModule(new SignalsModule());
 		}
 				
-		// add the module for link to link routing if enabled
-		if (config.controler().isLinkToLinkRoutingEnabled()){
-			controler.addOverridingModule(new LinkToLinkRoutingWithSignalsDataGuiceModule());
-		}
-		
 		controler.run();
 	}
